+++ source/OffsetGenerator.h
@@ -426,5 +426,64 @@ class OffsetGenRandomAlignedFullCoverage : public OffsetGenerator
         }
 };
 
+/**
+ * Offset generator for sequential strided access based on number of dataset threads.
+ */
+class OffsetGenStrided : public OffsetGenerator
+{
+    public:
+        OffsetGenStrided(uint64_t len, uint64_t offset, size_t blockSize,
+            size_t numDataSetThreads) :
+            numBytesTotal(len), numBytesLeft(len), startOffset(offset),
+            currentOffset(offset), blockSize(blockSize), numDataSetThreads(numDataSetThreads)
+        { }
+
+        virtual ~OffsetGenStrided() {}
+
+    protected:
+        uint64_t numBytesTotal;
+        uint64_t numBytesLeft;
+        uint64_t startOffset;
+        uint64_t currentOffset;
+        const size_t blockSize;
+        const size_t numDataSetThreads;
+
+    // inliners
+    public:
+        virtual void reset() override
+        {
+            numBytesLeft = numBytesTotal;
+            currentOffset = startOffset;
+        }
+
+        virtual void reset(uint64_t len, uint64_t offset) override
+        {
+            numBytesTotal = len;
+            numBytesLeft = len;
+            startOffset = offset;
+            currentOffset = offset;
+        }
+
+        virtual uint64_t getNextOffset() override
+            { return currentOffset; }
+
+        virtual size_t getBlockSize() const override
+            { return blockSize; }
+
+        virtual size_t getNextBlockSizeToSubmit() const override
+            { return std::min(numBytesLeft, blockSize); }
+
+        virtual uint64_t getNumBytesTotal() const override
+            { return numBytesTotal; }
+
+        virtual uint64_t getNumBytesLeftToSubmit() const override
+            { return numBytesLeft; }
+
+        virtual void addBytesSubmitted(size_t numBytes) override
+        {
+            numBytesLeft -= numBytes;
+            currentOffset += (blockSize * numDataSetThreads);
+        }
+};
 
 #endif /* OFFSETGENERATOR_H_ */